Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pillow
imagehash
exifread
opencv-python-headless
numpy
sqlalchemy
alembic
requests
python-dotenv
torch
torchvision
timm
albumentations
grad-cam
reportlab